restructured and added gitignore
