+++ .gitignore
@@ -1,0 +1,4 @@
+states.png
+/__pycache__/*
+*/__pycache__/*
+TODO
