+++ pyserial_docs_code/combination_serial_code_test.py
@@ -1,7 +1,7 @@
 import serial
 import time
 from serial_write_with_confirm import write_with_confirm
-from threaded_serial_reader2 import ThreadedSerialReader
+from threaded_serial_reader import ThreadedSerialReader
 
 if __name__ == '__main__':
     ser1 = serial.Serial(port="COM5", baudrate=9600, timeout=1)
