+++ pyserial_docs_code/threaded_serial_reader.py
